refactor: Filter in mobile should be a component
Move all related components to folder
